docker/agent/requirements.txt
python-dotenv
langchain
openai
requests
langchain-community
langgraph
langchain-openai
Flask
ipython
pyodbc
wikipedia